refactor(web): establish feature module foundation
