+++ apps/web/src/styles/base.css
@@ -1,0 +1,104 @@
+:root {
+  font-family: var(--font-sans);
+  color: var(--color-ink);
+  background: var(--color-canvas);
+  font-synthesis: none;
+}
+
+* {
+  box-sizing: border-box;
+}
+
+body {
+  margin: 0;
+  min-width: 320px;
+  min-height: 100vh;
+}
+
+a {
+  color: inherit;
+}
+
+button,
+.button {
+  border: 0;
+  border-radius: 999px;
+  background: var(--color-brand);
+  color: white;
+  font: inherit;
+  font-weight: 700;
+  padding: 0.78rem 1.2rem;
+  cursor: pointer;
+  text-decoration: none;
+  display: inline-block;
+}
+
+button:focus-visible,
+a:focus-visible,
+input:focus-visible,
+select:focus-visible,
+summary:focus-visible {
+  outline: 3px solid var(--color-focus);
+  outline-offset: 3px;
+}
+
+.secondary {
+  background: #dce9df;
+  color: #164c3a;
+  margin-left: 0.75rem;
+}
+
+input,
+select {
+  max-width: 100%;
+  min-height: 44px;
+  border: 1px solid #aebdb5;
+  border-radius: var(--radius-control);
+  padding: 0.65rem 0.75rem;
+  background: var(--color-surface);
+  color: var(--color-ink);
+  font: inherit;
+}
+
+dt {
+  color: var(--color-muted);
+}
+
+dd {
+  margin: 0;
+  overflow-wrap: anywhere;
+}
+
+pre {
+  white-space: pre-wrap;
+  overflow-wrap: anywhere;
+}
+
+code {
+  display: inline-block;
+  max-width: 100%;
+  padding: 0.55rem 0.7rem;
+  background: #eef2ec;
+  border-radius: 7px;
+  overflow-wrap: anywhere;
+}
+
+table {
+  width: 100%;
+  border-collapse: collapse;
+}
+
+th,
+td {
+  text-align: left;
+  padding: 0.9rem 1rem;
+  border-bottom: 1px solid var(--color-border-subtle);
+  white-space: nowrap;
+}
+
+th {
+  color: #617168;
+  font-size: 0.75rem;
+  text-transform: uppercase;
+  letter-spacing: 0.05em;
+}
