+++ apps/web/src/ui/empty-state.tsx
@@ -1,0 +1,10 @@
+import type { ReactNode } from "react";
+
+export function EmptyState({ title, children }: { title: string; children: ReactNode }) {
+  return (
+    <section className="empty-state">
+      <h2>{title}</h2>
+      <p>{children}</p>
+    </section>
+  );
+}
