+++ apps/web/src/ui/form-card.tsx
@@ -1,0 +1,9 @@
+import type { ComponentProps } from "react";
+
+export function FormCard({ children, ...props }: ComponentProps<"form">) {
+  return (
+    <form className="form-card" {...props}>
+      {children}
+    </form>
+  );
+}
