+++ apps/web/src/ui/page.tsx
@@ -1,0 +1,26 @@
+import type { ReactNode } from "react";
+
+export function Page({
+  title,
+  eyebrow,
+  action,
+  children,
+}: {
+  title: string;
+  eyebrow?: string;
+  action?: ReactNode;
+  children: ReactNode;
+}) {
+  return (
+    <>
+      <header className="page-header">
+        <div>
+          {eyebrow ? <span className="eyebrow">{eyebrow}</span> : null}
+          <h1>{title}</h1>
+        </div>
+        {action}
+      </header>
+      {children}
+    </>
+  );
+}
